Interrupt state when E depletion and load energy consumption
When entering a state, an event scheduling the computation of the energy is set
after the maximum time that the device can stay in that state before turning
off. If the device switches to a new state, the event is updated. Instead, if
that event occurs, the UpdateEnergySouce function is called, which checks if the
energy is depleted. Since errors occurred in the value comparison, a epsilon
value is used.

Also, a function to compute the *load* energy consumption when consuming a given
current for a given duration is added. This is used in the LoraRadioEnergyModel
to obtain the total energy consumption (of the load).